TFG-Backend/requirements.txt
asgiref==3.6.0
Django==4.2
sqlparse==0.4.4
typing_extensions==4.13.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==3.14.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
simplejson==3.19.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
